main router, rename newusers to createuser, put login into a file, moved new users to pages
