+++ frontend/src/pages/Login/Login.scss
@@ -1,10 +1,10 @@
-@import "../variables";
+@import "../../variables";
 
 .login__content {
   font-weight: bold;
   color: $primary-dark;
   font-family: $font-app;
-  background-image: url("../assets/image/background-image.png");
+  background-image: url("../../assets/image/background-image.png");
   background-repeat: no-repeat;
   background-size: cover;
   background-position: center;
